Scan2Topic-master/requirements.txt
attrs==19.3.0
backboned-unet==0.0.1
backcall==0.1.0
bleach==3.1.4
boto==2.49.0
boto3==1.14.33
botocore==1.17.33
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
cython==0.29.14
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
efficientnet-pytorch==0.6.3
entrypoints==0.3
ffmpeg==1.4
gensim==3.8.3
h5py==2.10.0
idna==2.9
imagecodecs==2020.2.18
imageio==2.8.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
jinja2==2.11.1
jmespath==0.10.0
joblib==0.14.1
json5==0.9.4
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-core==4.6.3
jupyterlab==2.1.0
jupyterlab-server==1.1.0
kiwisolver==1.2.0
markupsafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
munch==2.5.0
nbconvert==5.6.1
nbformat==5.0.5
networkx==2.4
notebook==6.0.3
opencv-contrib-python==4.2.0.34
pandas==1.0.3
pandocfilters==1.4.2
parso==0.6.2
pickleshare==0.7.5
pillow==7.2.0
pretrainedmodels==0.7.4
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.11.3
pycocotools==2.0
pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
pyspellchecker==0.5.4
python-dateutil==2.8.1
pytz==2020.1
pywavelets==1.1.1
pywin32==225
pywinpty==0.5.7
pyzmq==19.0.0
regex==2020.7.14
requests==2.23.0
s3transfer==0.3.3
scikit-image==0.17.2
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.1
segmentation-models-pytorch==0.1.0
send2trash==1.5.0
sklearn==0.0
smart-open==2.1.0
tensorboardx==2.0
terminado==0.8.3
testpath==0.4.4
tifffile==2020.5.11
tiffile==2018.10.18
torchsummary==1.5.1
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
urllib3==1.25.8
videofig==1.0.1
wcwidth==0.1.9
webencodings==0.5.1
widgetsnbextension==3.5.1